Refactor findSteamAppDetails
