+++ src/utils/gameUtils.ts
@@ -1,15 +1,5 @@
-import {
-  ActionRowBuilder,
-  ButtonBuilder,
-  ButtonStyle,
-  ChatInputCommandInteraction,
-  EmbedBuilder,
-  hyperlink,
-  Message,
-} from 'discord.js';
+import { ChatInputCommandInteraction, EmbedBuilder, hyperlink } from 'discord.js';
 import { isSteamAppDetail, SteamAppDetail } from '../SteamAppDetail';
-import { findSteamAppDetails } from '../services/steamService';
-import { MessageActionRowComponentBuilder } from '@discordjs/builders';
 import { GameModel } from '../configuration/models/game.model';
 import { CategoryModel } from '../configuration/models/category.model';
 import { hideLinkEmbed } from '@discordjs/formatters';
@@ -22,29 +12,6 @@ export const enum ButtonCustomIds {
   edit = 'edit',
   cancel = 'cancel',
 }
-
-const generateActionRow = (): ActionRowBuilder<MessageActionRowComponentBuilder> => {
-  const confirm = new ButtonBuilder()
-    .setCustomId(ButtonCustomIds.confirm)
-    .setLabel('Add')
-    .setStyle(ButtonStyle.Success);
-
-  const edit = new ButtonBuilder()
-    .setCustomId(ButtonCustomIds.edit)
-    .setLabel('Edit')
-    .setStyle(ButtonStyle.Primary);
-
-  const cancel = new ButtonBuilder()
-    .setCustomId(ButtonCustomIds.cancel)
-    .setLabel('Cancel')
-    .setStyle(ButtonStyle.Secondary);
-
-  return new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
-    cancel,
-    edit,
-    confirm,
-  );
-};
 
 export const generateGameEmbed = (content: GameModel | SteamAppDetail): EmbedBuilder => {
   const { name, steamAppid, genres = [] } = content;
@@ -93,29 +60,6 @@ export const generateGameEmbed = (content: GameModel | SteamAppDetail): EmbedBui
     )
     .setURL(steamAppid ? generateSteamAppUrl(steamAppid) : null)
     .setFooter({ text: `added by ${owner}` });
-};
-export const findAndDisplaySteamAppDetails = async (
-  interaction: ChatInputCommandInteraction,
-  query: string,
-  content = "Is this the game you'd like to add?",
-): Promise<[SteamAppDetail | undefined, Message]> => {
-  const details = await findSteamAppDetails(query);
-
-  if (!details) {
-    const message = await interaction.editReply({
-      content: `Sorry, I wasn't able to find "${query}" in the Steam store :sweat:.`,
-    });
-    return [details, message];
-  }
-
-  const embed = generateGameEmbed(details);
-  const row = generateActionRow();
-  const message = await interaction.editReply({
-    content,
-    embeds: [embed],
-    components: [row],
-  });
-  return [details, message];
 };
 
 export const generateSteamAppUrl = (
