+++ src/utils/steamUtils.ts
@@ -1,5 +1,15 @@
 import { SteamAppDetail } from '../SteamAppDetail';
-import { EmbedBuilder } from 'discord.js';
+import {
+  ActionRowBuilder,
+  ButtonBuilder,
+  ButtonStyle,
+  ChatInputCommandInteraction,
+  EmbedBuilder,
+  Message,
+} from 'discord.js';
+import { findSteamAppDetails } from '../services/steamService';
+import { ButtonCustomIds, generateGameEmbed } from './gameUtils';
+import { MessageActionRowComponentBuilder } from '@discordjs/builders';
 
 export const generateSteamAppEmbed = ({
   name,
@@ -21,4 +31,52 @@ export const generateSteamAppEmbed = ({
       { name: 'Price', value: finalFormatted, inline: true },
     )
     .setURL(website);
+};
+
+const generateCreationActionRow =
+  (): ActionRowBuilder<MessageActionRowComponentBuilder> => {
+    const confirm = new ButtonBuilder()
+      .setCustomId(ButtonCustomIds.confirm)
+      .setLabel('Add')
+      .setStyle(ButtonStyle.Success);
+
+    const edit = new ButtonBuilder()
+      .setCustomId(ButtonCustomIds.edit)
+      .setLabel('Edit')
+      .setStyle(ButtonStyle.Primary);
+
+    const cancel = new ButtonBuilder()
+      .setCustomId(ButtonCustomIds.cancel)
+      .setLabel('Cancel')
+      .setStyle(ButtonStyle.Secondary);
+
+    return new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+      cancel,
+      edit,
+      confirm,
+    );
+  };
+
+export const findAndDisplaySteamAppDetails = async (
+  interaction: ChatInputCommandInteraction,
+  query: string,
+  content = "Is this the game you'd like to add?",
+): Promise<[SteamAppDetail | undefined, Message]> => {
+  const details = await findSteamAppDetails(query);
+
+  if (!details) {
+    const message = await interaction.editReply({
+      content: `Sorry, I wasn't able to find "${query}" in the Steam store :sweat:.`,
+    });
+    return [details, message];
+  }
+
+  const embed = generateGameEmbed(details);
+  const row = generateCreationActionRow();
+  const message = await interaction.editReply({
+    content,
+    embeds: [embed],
+    components: [row],
+  });
+  return [details, message];
 };
